+++ src/bvals/cc/bvals_cc.cpp
@@ -62,8 +62,16 @@ CellCenteredBoundaryVariable::CellCenteredBoundaryVariable(std::weak_ptr<MeshBlo
 
   InitBoundaryData(bd_var_, BoundaryQuantity::cc);
 
+#ifdef MPI_PARALLEL
+  cc_var_comm = pmy_mesh_->GetMPIComm(label);
+#endif
+
   if (pmy_mesh_->multilevel) { // SMR or AMR
     InitBoundaryData(bd_var_flcor_, BoundaryQuantity::cc_flcor);
+
+#ifdef MPI_PARALLEL
+    cc_flcor_comm = pmy_mesh_->GetMPIComm(label + "_flcor");
+#endif
   }
 }
 
@@ -140,8 +148,6 @@ int CellCenteredBoundaryVariable::ComputeFluxCorrectionBufferSize(
 void CellCenteredBoundaryVariable::SetupPersistentMPI() {
 #ifdef MPI_PARALLEL
   std::shared_ptr<MeshBlock> pmb = GetBlockPointer();
-  // TODO(PG) move to constructor once MPI_Comm_dup is in Mesh constructor
-  cc_var_comm = pmb->pmy_mesh->mpi_comm_map[label()];
   int &mylevel = pmb->loc.level;
 
   int ssize, rsize;
@@ -189,8 +195,6 @@ void CellCenteredBoundaryVariable::SetupPersistentMPI() {
                                         &(bd_var_.req_recv[nb.bufid])));
 
       if (pmy_mesh_->multilevel && nb.ni.type == NeighborConnect::face) {
-        // TODO(PG) move to constructor once MPI_Comm_dup is in Mesh constructor
-        cc_flcor_comm = pmb->pmy_mesh->mpi_comm_map[label() + "_flcor"];
         // TODO(JL): could we call ComputeFluxCorrectionBufferSize here to reduce code
         // duplication?
         int size;