+++ src/interface/swarm.cpp
@@ -860,8 +860,6 @@ void Swarm::SetNeighborIndices3D_() {
 
 void Swarm::SetupPersistentMPI() {
   auto pmb = GetBlockPointer();
-  // TODO(PG) move to constructor once MPI_Comm_dup is in Mesh constructor
-  vbswarm->SetComm(pmb->pmy_mesh->mpi_comm_map[label()]);
   vbswarm->SetupPersistentMPI();
 
   const int ndim = pmb->pmy_mesh->ndim;
@@ -1215,7 +1213,7 @@ void Swarm::AllocateComms(std::weak_ptr<MeshBlock> wpmb) {
   std::shared_ptr<MeshBlock> pmb = wpmb.lock();
 
   // Create the boundary object
-  vbswarm = std::make_shared<BoundarySwarm>(pmb);
+  vbswarm = std::make_shared<BoundarySwarm>(pmb, label_);
 
   // Enroll SwarmVariable object
   vbswarm->bswarm_index = pmb->pbswarm->bswarms.size();